Issue 149: Move exception interceptor into one class. (shv)
